feat: 빌드오류 suspense로 개선, 카드 타이틀 길면 말줌임 추가
